+++ resources/views/inc/introduction.blade.php
@@ -1,0 +1,5 @@
+<div id="introduction" class="mt-2">
+    <h3>#Introduction</h3>
+   <p class="lead"> Lorem ipsum dolor sit amet consectetur, adipisicing elit. Neque excepturi mollitia error voluptas sint similique, assumenda aliquid quod eius! Necessitatibus eveniet ipsa optio consequatur velit, nostrum officiis veniam eum quae eos eius quam, omnis dolor temporibus dolorem aspernatur voluptate, dicta tempora natus ratione suscipit facere illum ad! Similique, sint autem?
+    voluptas sint similique, assumenda aliquid quod eius! Necessitatibus eveniet ipsa optio consequatur velit, nostrum officiis veniam eum quae eos eius quam, omnis dolor temporibus dolorem aspernatur voluptate, dicta tempora natus ratione suscipit facere illum ad! Similique, sint autem?</p>
+</div>
