+++ resources/views/inc/sidebar.blade.php
@@ -1,0 +1,12 @@
+<div class="col-md-3 bg-light" style="min-height: 100vh">
+    <div class="text-center mt-5">
+     <img src="./talk-to-api.png" alt="" class="">
+    </div>
+    <hr>
+     <a class="btn btn-link text-left w-100" href="/">
+         Getting Started
+     </a>
+     <a class="btn btn-link text-left w-100" href="todo">
+        To Do Api
+    </a>
+ </div> <!-- col-3 -->
